+++ tests/integration/adapter/clickhouse/test_clickhouse_errors.py
