+++ tests/integration/adapter/clickhouse/test_clickhouse_s3.py
