+++ tests/integration/adapter/clickhouse/test_clickhouse_source_schema.py
