+++ tests/integration/adapter/clickhouse/test_clickhouse_upper_case.py
