+++ tests/integration/adapter/dbt_show/test_dbt_show.py
@@ -1,0 +1,9 @@
+from dbt.tests.adapter.dbt_show.test_dbt_show import BaseShowLimit, BaseShowSqlHeader
+
+
+class TestShowLimit(BaseShowLimit):
+    pass
+
+
+class TestShowSqlHeader(BaseShowSqlHeader):
+    pass
