+++ tests/integration/adapter/grants/test_distributed_grants.py
@@ -1,11 +1,7 @@
 import os
 
 import pytest
-from dbt.tests.adapter.grants.test_incremental_grants import BaseIncrementalGrants
-from dbt.tests.adapter.grants.test_invalid_grants import BaseInvalidGrants
 from dbt.tests.adapter.grants.test_model_grants import BaseModelGrants
-from dbt.tests.adapter.grants.test_seed_grants import BaseSeedGrants
-from dbt.tests.adapter.grants.test_snapshot_grants import BaseSnapshotGrants
 from dbt.tests.util import get_manifest, run_dbt_and_capture, write_file
 
 distributed_table_model_schema_yml = """
@@ -18,31 +14,6 @@ models:
         select: ["{{ env_var('DBT_TEST_USER_1') }}"]
         insert: ["{{ env_var('DBT_TEST_USER_2') }}"]
 """
-
-
-class TestModelGrants(BaseModelGrants):
-    pass
-
-
-class TestIncrementalGrants(BaseIncrementalGrants):
-    pass
-
-
-class TestSeedGrants(BaseSeedGrants):
-    pass
-
-
-class TestInvalidGrants(BaseInvalidGrants):
-    def grantee_does_not_exist_error(self):
-        return "511"
-
-    # ClickHouse doesn't give a very specific error for an invalid privilege
-    def privilege_does_not_exist_error(self):
-        return "Syntax error"
-
-
-class TestSnapshotGrants(BaseSnapshotGrants):
-    pass
 
 
 class TestDistributedTableModelGrants(BaseModelGrants):
