+++ tests/integration/adapter/grants/test_incremental_grants.py
@@ -1,0 +1,5 @@
+from dbt.tests.adapter.grants.test_incremental_grants import BaseIncrementalGrants
+
+
+class TestIncrementalGrants(BaseIncrementalGrants):
+    pass
