+++ tests/integration/adapter/grants/test_model_grants.py
@@ -1,0 +1,5 @@
+from dbt.tests.adapter.grants.test_model_grants import BaseModelGrants
+
+
+class TestModelGrants(BaseModelGrants):
+    pass
