+++ tests/integration/adapter/grants/test_seed_grants.py
@@ -1,0 +1,5 @@
+from dbt.tests.adapter.grants.test_seed_grants import BaseSeedGrants
+
+
+class TestSeedGrants(BaseSeedGrants):
+    pass
