+++ tests/integration/adapter/grants/test_snapshot_grants.py
@@ -1,0 +1,5 @@
+from dbt.tests.adapter.grants.test_snapshot_grants import BaseSnapshotGrants
+
+
+class TestSnapshotGrants(BaseSnapshotGrants):
+    pass
