+++ tests/integration/adapter/query_comment/test_query_comment.py
