+++ tests/integration/adapter/relations/test_changing_relation_type.py
